// src/app/shared/models/entities/member.ts
export interface Member {
    Active: number;
    Address: string;
    AllowCalendar: number;
    Avatar: {[key: string]: string};
    City: string;
    Company: string;
    Country: string;
    CreatePersonnelFile: number;
    Created: string;
    CurrentDataStore: {[key: string]: any};
    DataStores: string[];
    Email: string;
    FirstName: string;
    Groups: any[];
    Icon: string;
    LastEdited: string;
    Phone: string;
    PostCode: string;
    Slug: string;
    Surname: string;
    avatarURL: string;
    Settings: any;
}